requirements.txt
# ── Web Framework ──
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary>=2.9.9

# ── Cache / Real-time ──
redis==5.0.4
aioredis==2.0.1

# ── Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Task Scheduler ──
apscheduler==3.10.4

# ── AI ──
anthropic>=0.40.0

# ── Telegram ──
python-telegram-bot==21.1.1

# ── Market Data ──
smartapi-python>=1.3.4
pyotp>=2.9.0
websocket-client>=1.6.0
httpx==0.27.0
aiohttp==3.9.5
feedparser==6.0.11

# ── Config & Env ──
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Utilities ──
pytz==2024.1
arrow==1.3.0
tenacity==8.3.0
structlog==24.2.0

# ── Testing ──
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
pytest-mock==3.14.0
factory-boy==3.3.0
playwright==1.44.0
aiosqlite==0.20.0
